+++ packages/@webex/plugin-meetings/src/media/index.ts
@@ -3,13 +3,8 @@
  */
 /* globals navigator */
 
-import {
-  LocalCameraTrack,
-  LocalDisplayTrack,
-  LocalMicrophoneTrack,
-  RoapMediaConnection,
-  MultistreamRoapMediaConnection,
-} from '@webex/internal-media-core';
+import {RoapMediaConnection, MultistreamRoapMediaConnection} from '@webex/internal-media-core';
+import {LocalCameraTrack, LocalDisplayTrack, LocalMicrophoneTrack} from '@webex/media-helpers';
 import LoggerProxy from '../common/logs/logger-proxy';
 import {AUDIO_INPUT, VIDEO_INPUT, MEDIA_TRACK_CONSTRAINT} from '../constants';
 import Config from '../config';
@@ -18,6 +13,8 @@ import MediaError from '../common/errors/media';
 import BrowserDetection from '../common/browser-detection';
 
 const {isBrowser} = BrowserDetection();
+
+type MultistreamConnectionConfig = ConstructorParameters<typeof MultistreamRoapMediaConnection>[0];
 
 export type BundlePolicy = ConstructorParameters<
   typeof MultistreamRoapMediaConnection
@@ -166,17 +163,19 @@ Media.createMediaConnection = (
   }
 
   if (isMultistream) {
-    return new MultistreamRoapMediaConnection(
-      {
-        iceServers,
-        enableMainAudio:
-          mediaProperties.mediaDirection?.sendAudio || mediaProperties.mediaDirection?.receiveAudio,
-        enableMainVideo:
-          mediaProperties.mediaDirection?.sendVideo || mediaProperties.mediaDirection?.receiveVideo,
-        bundlePolicy,
-      },
-      debugId
-    );
+    const config: MultistreamConnectionConfig = {
+      iceServers,
+      enableMainAudio:
+        mediaProperties.mediaDirection?.sendAudio || mediaProperties.mediaDirection?.receiveAudio,
+      enableMainVideo:
+        mediaProperties.mediaDirection?.sendVideo || mediaProperties.mediaDirection?.receiveVideo,
+    };
+
+    if (bundlePolicy) {
+      config.bundlePolicy = bundlePolicy;
+    }
+
+    return new MultistreamRoapMediaConnection(config, debugId);
   }
 
   if (!mediaProperties) {