+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.0.0-beta.88](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.87...v3.0.0-beta.88) (2023-04-28)
+
+
+### Bug Fixes
+
+* **breakout:** add editlock to assign api - SPARK-425241 ([#2780](https://github.com/webex/webex-js-sdk/issues/2780)) ([75a3300](https://github.com/webex/webex-js-sdk/commit/75a33001ad0fe880cce85889ff3f86ac7e71fb6a))
+
+## [3.0.0-beta.87](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.86...v3.0.0-beta.87) (2023-04-27)
+
+
+### Features
+
+* change of SDK multistream APIs to use webrtc-core local tracks ([#2778](https://github.com/webex/webex-js-sdk/issues/2778)) ([a7521a3](https://github.com/webex/webex-js-sdk/commit/a7521a3f1025af481af39f7e08481af8daa67ce2))
+
+## [3.0.0-beta.86](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.85...v3.0.0-beta.86) (2023-04-26)
+
+
+### Bug Fixes
+
+* default bundle policy from WCME not applied ([#2785](https://github.com/webex/webex-js-sdk/issues/2785)) ([9a313e1](https://github.com/webex/webex-js-sdk/commit/9a313e115ed59d2baacb7e4b4c935c142136c5e0))
+
 ## [3.0.0-beta.85](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.84...v3.0.0-beta.85) (2023-04-26)